Fixed recording bug
Fixed a bug where after a recorded game, if the player chose to exit the game, the Move Log.txt file was not being cleared of the previous game's moves. This caused the next recorded game's moves to be appended onto the file along with the previous game's moves. This has rectified so the Move Log.txt file is cleared and reset when a player chooses to exit the program without replaying the previous game